agents/requirements.txt
psycopg[binary]
python-dotenv
google-genai
google-adk
